selfupdate updated but not tested.

+++ wasanbon/core/management/admin/selfupdate.py
@@ -5,52 +5,68 @@ import os, sys, subprocess, shutil
 from wasanbon import util
 
 
-def pull_and_update(verbose):
+def pull_and_update(verbose, force):
     cwd = os.getcwd()
     os.chdir(os.path.join(wasanbon.rtm_temp, 'wasanbon'))
 
     if verbose:
         sys.stdout.write(' - Pulling from %s\n' % wasanbon.setting['common']['repository']['wasanbon']['git'])
-        pass
 
     cmd = [wasanbon.setting['local']['git'], 'pull']
     p = subprocess.Popen(cmd, stdout=subprocess.PIPE)
     output = p.stdout.readline()
     if output.strip() == 'Already up-to-date.':
         sys.stdout.write(output)
-        return
+        return False
 
-    install(verbose)
+    return install(verbose=verbose, force=force)
 
-def clone_and_update(verbose):
+def clone_and_update(verbose, force):
+    cwd = os.getcwd()
+    os.chdir(wasanbon.rtm_temp)
+
     if verbose: 
         sys.stdout.write(' - Cloning %s\n' % wasanbon.setting['common']['repository']['wasanbon']['git'])
-        pass
 
-    cmd = [wasanbon.setting['local']['git'], 
-           'clone', 
+    cmd = [wasanbon.setting['local']['git'], 'clone', 
            wasanbon.setting['common']['repository']['wasanbon']['git']]
     subprocess.call(cmd)
-    os.chdir('wasanbon')
-    pass
+
+    return install(verbose=verbose, force=force)
+
 
 def cleanup(verbose):
-    cwd = os.getcwd()
-    os.chdir(os.path.join(wasanbon.rtm_temp, 'wasanbon'))
+    dirname = os.path.join(wasanbon.rtm_temp, 'wasanbon')
+    if not os.path.isdir(dirname):
+        return False
+
     if verbose:
         sys.stdout.write(' - Cleanly uninstalled.\n')
-    shutil.rmtree(wasanbon.__path__[0])
-    shutil.rmtree('build')
-    os.chdir(cwd)
-    pass
-
-def install(verbose):
+    
     cwd = os.getcwd()
-    os.chdir(os.path.join(wasanbon.rtm_temp, 'wasanbon'))
+    os.chdir(dirname)
+    shutil.rmtree(wasanbon.__path__[0])
+    
+    os.chdir(cwd)
+    return True
+
+def install(verbose, force):
+    dirname = os.path.join(wasanbon.rtm_temp, 'wasanbon')
+    if not os.path.isdir(dirname):
+        return False
+    cwd = os.getcwd()
+    if verbose:
+        sys.stdout.write(' - Installing wasanbon from %s.\n' % dirname)
+
+    if force:
+        cleanup(verbose)
+
+    shutil.rmtree('build')
+    os.chdir(dirname)
     cmd = ['python', 'setup.py', 'install']
     subprocess.call(cmd)
     os.chdir(cwd)
-    pass
+    return True
 
 class Command(object):
     def __init__(self):
@@ -58,15 +74,17 @@ class Command(object):
 
     def execute_with_argv(self, argv, verbose=False, force=False, clean=False):
         cwd = os.getcwd()
+
         if verbose:
             sys.stdout.write(' - Changing directory to %s\n' % wasanbon.rtm_temp)
         os.chdir(wasanbon.rtm_temp)
-
         if not os.path.isdir('wasanbon'):
+            if not clean:
+                clone_and_update(verbose=verbose, force=force)
         else:
+            if clean:
+                cleanup(verbose=verbose)
+            else:
+                pull_and_update(verbose=verbose, force=force)
 
-        if clean:
-            cleanup(verbose)
-        else:
-            install(verbose)
         os.chdir(cwd)
